refactor(appium): reorganize bootstrap entities into a separate subfolder
