+++ packages/appium/lib/schema/index.ts
@@ -1,2 +1,4 @@
 export * from './schema';
+export * from './format-errors';
 export * from './cli-args';
+export * from './cli-args-guards';
