+++ packages/appium/test/e2e/config-file.e2e.spec.ts
@@ -1,5 +1,5 @@
 import {DRIVER_TYPE} from '../../lib/constants';
-import {readConfigFile} from '../../lib/config-file';
+import {readConfigFile} from '../../lib/bootstrap/config-file';
 import {finalizeSchema, registerSchema, resetSchema} from '../../lib/schema/schema';
 import extSchema from '../fixtures/driver-schema';
 import {resolveFixture} from '../helpers';
