+++ packages/appium/test/helpers.ts
@@ -1,7 +1,7 @@
 import getPort from 'get-port';
 import path from 'node:path';
 import rewiremock, {addPlugin, overrideEntryPoint, plugins} from 'rewiremock';
-import {insertAppiumPrefixes} from '../lib/utils';
+import {insertAppiumPrefixes} from '../lib/helpers/capability';
 
 const TEST_HOST = '127.0.0.1';
