--- packages/appium/test/unit/config.spec.ts
@@ -1,277 +1,0 @@
-import _ from 'lodash';
-import chai from 'chai';
-import chaiAsPromised from 'chai-as-promised';
-import {promises as fs} from 'node:fs';
-import os from 'node:os';
-import path from 'node:path';
-import type {SinonSandbox, SinonSpy} from 'sinon';
-import {createSandbox} from 'sinon';
-import {getParser} from '../../lib/cli/parser';
-import {
-  checkNodeOk,
-  getBuildInfo,
-  getNonDefaultServerArgs,
-  showBuildInfo,
-  showConfig,
-  requireDir,
-} from '../../lib/config';
-import {PLUGIN_TYPE} from '../../lib/constants';
-import {
-  finalizeSchema,
-  getDefaultsForSchema,
-  registerSchema,
-  resetSchema,
-} from '../../lib/schema/schema';
-
-const {expect} = chai;
-chai.use(chaiAsPromised);
-
-describe('Config', function () {
-  let sandbox: SinonSandbox;
-
-  beforeEach(function () {
-    sandbox = createSandbox();
-  });
-
-  afterEach(function () {
-    sandbox.restore();
-  });
-
-  describe('Appium config', function () {
-    let log: SinonSpy;
-    let dir: SinonSpy;
-    beforeEach(function () {
-      log = sandbox.spy(console, 'log');
-      dir = sandbox.spy(console, 'dir');
-    });
-
-    describe('showBuildInfo()', function () {
-      it('should log build info to console', async function () {
-        const config = getBuildInfo();
-        await showBuildInfo();
-        expect(log.calledOnce).to.be.true;
-        expect(log.firstCall.args).to.contain(JSON.stringify(config));
-      });
-    });
-
-    describe('showConfig()', function () {
-      describe('when a config file is present', function () {
-        it('should dump the current Appium config', function () {
-          showConfig(
-            {address: 'bar'},
-            {
-              config: {
-                // @ts-expect-error
-                server: {'callback-address': 'quux'},
-              },
-            },
-            {port: 1234},
-            {allowCors: false}
-          );
-          expect(log.calledWith('Appium Configuration\n')).to.be.true;
-        });
-
-        it('should skip empty objects', function () {
-          showConfig(
-            // @ts-expect-error
-            {foo: 'bar', cows: {}, pigs: [], sheep: 0, ducks: false},
-            {config: {server: {address: 'quux'}}},
-            {spam: 'food'},
-            {}
-          );
-            expect(dir.calledWith({foo: 'bar', sheep: 0, ducks: false})).to.be.true;
-        });
-      });
-
-      describe('when a config file is not present', function () {
-        it('should dump the current Appium config (sans config file contents)', function () {
-          showConfig(
-            // @ts-expect-error
-            {foo: 'bar', cows: {}, pigs: [], sheep: 0, ducks: false},
-            {},
-            {spam: 'food'},
-            {}
-          );
-            expect(log.calledWith('\n(no configuration file loaded)')).to.be.true;
-        });
-      });
-
-      describe('when no CLI arguments (other than --show-config) provided', function () {
-        it('should not dump CLI args', function () {
-          showConfig({}, {}, {}, {});
-          expect(log.calledWith('\n(no CLI parameters provided)')).to.be.true;
-        });
-      });
-    });
-  });
-
-  describe('node.js config', function () {
-    const _process = process;
-    before(function () {
-      // need to be able to write to process.version
-      // but also to have access to process methods
-      // so copy them over to a writable object
-      process = {...process}; // eslint-disable-line no-global-assign
-    });
-    after(function () {
-      process = _process; // eslint-disable-line no-global-assign
-    });
-    describe('checkNodeOk', function () {
-      describe('unsupported nodes', function () {
-        const unsupportedVersions = [
-          'v0.1',
-          'v0.9.12',
-          'v0.10.36',
-          'v0.12.14',
-          'v4.4.7',
-          'v5.7.0',
-          'v6.3.1',
-          'v7.1.1',
-          'v8.0.0',
-          'v9.2.3',
-          'v10.1.0',
-          'v11.0.0',
-          'v12.0.0',
-          'v14.0.0',
-          'v14.17.0',
-          'v14.17.5',
-          'v16.0.0',
-          'v20.18.0',
-          'v22.10.0',
-        ];
-        for (const version of unsupportedVersions) {
-          it(`should fail if node is ${version}`, function () {
-            // @ts-expect-error
-            process.version = version;
-            expect(checkNodeOk).to.throw();
-          });
-        }
-      });
-
-      describe('supported nodes', function () {
-        it('should succeed if node is ^20.19.0', function () {
-          // @ts-expect-error
-          process.version = 'v20.19.0';
-          expect(checkNodeOk).to.not.throw();
-          // @ts-expect-error
-          process.version = 'v20.100.0';
-          expect(checkNodeOk).to.not.throw();
-        });
-        it('should succeed if node is 22.12+', function () {
-          // @ts-expect-error
-          process.version = 'v22.12.0';
-          expect(checkNodeOk).to.not.throw();
-          // @ts-expect-error
-          process.version = 'v100.0.0';
-          expect(checkNodeOk).to.not.throw();
-        });
-      });
-    });
-  });
-
-  describe('server arguments', function () {
-    let args;
-
-    describe('getNonDefaultServerArgs', function () {
-      describe('without extension schemas', function () {
-        beforeEach(function () {
-          resetSchema();
-          getParser(true);
-          // get all the defaults
-          args = getDefaultsForSchema();
-        });
-        it('should show none if we have all the defaults', function () {
-          const nonDefaultArgs = getNonDefaultServerArgs(args);
-          expect(nonDefaultArgs).to.be.empty;
-        });
-        it('should catch a non-default argument', function () {
-          args.allowCors = true;
-          const nonDefaultArgs = getNonDefaultServerArgs(args);
-          expect(nonDefaultArgs).to.eql({allowCors: true});
-        });
-        describe('when arg is an array', function () {
-          it('should return the arg as an array', function () {
-            args.usePlugins = ['all'];
-            expect(getNonDefaultServerArgs(args)).to.eql({usePlugins: ['all']});
-          });
-        });
-      });
-      describe('with extension schemas', function () {
-        beforeEach(function () {
-          resetSchema();
-          registerSchema(PLUGIN_TYPE, 'crypto-fiend', {
-            type: 'object',
-            properties: {elite: {type: 'boolean', default: true}},
-          });
-          finalizeSchema();
-          getParser(true);
-          args = getDefaultsForSchema();
-        });
-
-        it('should take extension schemas into account', function () {
-          const nonDefaultArgs = getNonDefaultServerArgs(args);
-          expect(nonDefaultArgs).to.be.empty;
-        });
-
-        it('should catch a non-default argument', function () {
-          args['plugin.crypto-fiend.elite'] = false;
-          const nonDefaultArgs = getNonDefaultServerArgs(args);
-          expect(nonDefaultArgs).to.eql(_.set({}, 'plugin.crypto-fiend.elite', false));
-        });
-      });
-    });
-  });
-
-  describe('requireDir', function () {
-    it('should fail to use a dir with incorrect permissions', async function () {
-      await expect(requireDir('/private/if_you_run_with_sudo_this_wont_fail')).to.be.rejectedWith(
-        /must exist/
-      );
-    });
-    it('should fail to use an undefined dir', async function () {
-      // @ts-expect-error
-      await expect(requireDir()).to.be.rejectedWith(/must exist/);
-    });
-    it('should fail to use a non-writeable dir', async function () {
-      const tempDir = await fs.mkdtemp(path.join(os.tmpdir(), 'appium-requireDir-test-'));
-      try {
-        await fs.chmod(tempDir, 0o444);
-        await expect(requireDir(tempDir)).to.be.rejectedWith(/must be writeable/);
-      } finally {
-        await fs.chmod(tempDir, 0o700);
-        await fs.rmdir(tempDir);
-      }
-    });
-    it('should be able to use a dir with correct permissions', async function () {
-      await expect(requireDir('/tmp/test_tmp_dir/with/any/number/of/levels')).to.not.be.rejected;
-    });
-  });
-
-  describe('parsing args with empty argv[1]', function () {
-    let argv1;
-
-    before(function () {
-      argv1 = process.argv[1];
-    });
-
-    beforeEach(function () {
-      resetSchema();
-    });
-
-    after(function () {
-      process.argv[1] = argv1;
-    });
-
-    it('should not fail if process.argv[1] is undefined', function () {
-      process.argv[1] = '';
-      const args = getParser();
-      expect(args.prog).to.equal('appium');
-    });
-
-    it('should set "prog" to process.argv[1]', function () {
-      process.argv[1] = 'Hello World';
-      const args = getParser();
-      expect(args.prog).to.equal('Hello World');
-    });
-  });
-});
