+++ packages/appium/test/unit/extension/manifest.spec.ts
@@ -7,7 +7,7 @@ import type {SinonSandbox} from 'sinon';
 import {DRIVER_TYPE, PLUGIN_TYPE} from '../../../lib/constants';
 import {resolveFixture, rewiremock} from '../../helpers';
 import {initMocks} from './mocks';
-import {APPIUM_VER} from '../../../lib/config';
+import {APPIUM_VER} from '../../../lib/helpers/build';
 import EventEmitter from 'node:events';
 import type {MockAppiumSupport, MockGlob, MockPackageChanged} from './mocks';
